+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `disableActionSave` prop to `Datagrid`
 * Combine error and warning status for cells with 'error' type prioritized in `Datagrid`
 * Add grey corner to edited cells of `Datagrid`
+* Add navigation between cells of `Datagrid` with arrow keys in editing mode
 
 ## 4.1.3
 * fix to datagrid's scroll to row functionality